changed price to float64

+++ database/models/order.go
@@ -21,8 +21,8 @@ type Order struct {
 	Status string // filled, waiting or canceled
 	Amount int64  // amount of stocks in the order
 
-	Price 			int64
-	StopLossPrice   int64
+	Price         float64
+	StopLossPrice float64
 
 	TrailingPercentage float64
